feat: add a development server ping button
